Submessage indexing (#74)
* better at dumping transactions

* CP** - additional sql debugging

* CP** - test running but mocks incorrect

* CP - added test_schema_indexer.rs

* R - mostly code cleanup

* Fgit status - adds ids for variant submessage enum types but incorrect relation fields

* B - fixed typo in test struct

* CP!! - checkpointing around sub messages

* CPgit status - changing relational structure for sub messages but mapper is not complete

* T! Tests pass

* R! Cleaned up println statements, deleted comments

* T!! Tests sub message population (fails)

* T!! - Adds failing test for mapping submessages

* Fgit status stubbing out how mappers hold sub message schema refs

* F!! added member for tracking subschemas by name; member for mapping schemas

* F!! sub messages handled as a specialization of the regular recursive case

* R!! Ensure primary id method

* R!! uses foreign_key consistently; ensures a primary ID column on all tables

* t - fixed tests to expect id field

* t - fixed sql test for simple message

* t - test for SimpleMessage actually tests insert; refactored out compare and assert for transactions

* submessage mapping handled in db_mapper test

* T - submessage persistence test passes

* f - cargo fmt

* f - removed some kruft

* E - Updated clippy cleanup

+++ src/historical_parser.rs
@@ -343,7 +343,7 @@ pub async fn block_synchronizer(
         } else if remaining < config.block_page_size as i64 {
             current_height += remaining as u64;
         } else {
-            current_height += config.block_page_size as u64;
+            current_height += config.block_page_size;
         }
         outstanding_requests += 1;
         if outstanding_requests == max_requests {
